Steam: Disable using persistent connection for PollRequest as it causes force disconnects (thanks Vulpix for tip)
It might be reenabled in future again, we just need to make sure we're updating and cancelling it correctly in case we receive error or something. Now we remember the new response->nlc only in one branch of that if, which is obviously wrong, because we're trying to use the old nlc anyway without resetting it.

But maybe Steam servers doesn't support the polling thread with persistent connection at all, it needs to be tested further.

git-svn-id: http://svn.miranda-ng.org/main/trunk@15762 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c