SkypeWeb: Show better names for chat contacts
Based on idea of olegtarasov - https://github.com/miranda-ng/miranda-ng/pull/8

It primarily uses custom nicks (changed via chatroom), then it tries to get custom name of the contact in our contact list, or his nick (if custom name is not present), or as last option his id.

Also custom name is used on more places (when adding new contact, or kicking contact - and also the "initiator" of the kick).

git-svn-id: http://svn.miranda-ng.org/main/trunk@17211 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c